app/requirements.txt
flask
requests
python-dotenv
pyjwt
cryptography